+++ tests/testthat/test-tree_numbering.R
@@ -67,19 +67,24 @@ test_that("Replacement reorder functions work correctly", {
                RenumberEdges(edge[, 1], edge[, 2]))
 })
 
+test_that("Preorder handles large trees safely", {
+  Preorder(PectinateTree(23170))
+  expect_error(Preorder(RandomTree(23171)))
+})
+
 test_that("Preorder handles malformed trees without crashing", {
   treeDoubleNode <- read.tree(text = "((((((1,2)),3),4),5),6);")
   treePolytomy   <- read.tree(text = "((((1,2,3),4),5),6);")
   treeDoublyPoly <- read.tree(text = "(((((1,2,3)),4),5),6);")
-  
+
   reordered <- Preorder(treeDoubleNode)$edge
   expect_equal(11L, dim(reordered)[1])
   expect_equal(5L, sum(table(reordered[, 1]) == 2L))
-  
+
   reordered <- Preorder(treePolytomy)$edge
   expect_equal(9L, dim(reordered)[1])
   expect_equal(c(2L, 2L, 2L, 3L), as.integer(table(reordered[, 1])))
-  
+
   reordered <- Preorder(treeDoublyPoly)$edge
   expect_equal(10L, dim(reordered)[1])
   expect_equal(c(2L, 2L, 2L, 1L, 3L), as.integer(table(reordered[, 1])))